Fix raster.json and layer controller bugs, with disable3d fix (#1002)
* Refactor view config upgrade functions and validation functions

* Changelog

* Update tests

* Fix tests again

* Change variable name
