+++ src/app/CallbackPublisher.js
@@ -1,16 +1,12 @@
-import Ajv from 'ajv';
 import { useEffect } from 'react';
-import configSchema from '../schemas/config-1.0.1.schema.json';
-import cellSetsSchema from '../schemas/cell-sets.schema.json';
-import rasterSchema from '../schemas/raster.schema.json';
+import { SCHEMA_HANDLERS, LATEST_VERSION } from './view-config-versions';
 import { useViewConfigStore, useLoaders, useWarning } from './state/hooks';
 
 function validateViewConfig(viewConfig) {
   // Need the try-catch here since Zustand will actually
   // just catch and ignore errors in its subscription callbacks.
   try {
-    const validate = new Ajv()
-      .addSchema(cellSetsSchema).addSchema(rasterSchema).compile(configSchema);
+    const validate = SCHEMA_HANDLERS[LATEST_VERSION][0];
     const valid = validate(viewConfig);
     if (!valid) {
       const failureReason = JSON.stringify(validate.errors, null, 2);
