+++ src/app/Vitessce.js
@@ -1,24 +1,17 @@
 /* eslint-disable camelcase */
 import React, { useEffect, useMemo } from 'react';
-import Ajv from 'ajv';
 import {
   ThemeProvider, StylesProvider,
   createGenerateClassName,
 } from '@material-ui/core/styles';
 import isEqual from 'lodash/isEqual';
-import packageJson from '../../package.json';
 import { muiTheme } from '../components/shared-mui/styles';
-import configSchema from '../schemas/config-1.0.1.schema.json';
-import configSchema1_0_0 from '../schemas/config-1.0.0.schema.json';
-import configSchema0_1_0 from '../schemas/config-0.1.0.schema.json';
-import cellSetsSchema from '../schemas/cell-sets.schema.json';
-import rasterSchema from '../schemas/raster.schema.json';
 
 import VitessceGrid from './VitessceGrid';
 import Warning from './Warning';
 import CallbackPublisher from './CallbackPublisher';
 import { getComponent } from './component-registry';
-import { initialize, upgrade } from './view-config-utils';
+import { initialize, upgradeAndValidate } from './view-config-utils';
 
 const generateClassName = createGenerateClassName({
   disableGlobal: true,
@@ -76,55 +69,13 @@ export default function Vitessce(props) {
       }, false];
     }
     // Check if this is a "legacy" view config.
-    let upgradedConfig = config;
-    if (upgradedConfig.version === '0.1.0') {
-      // Validate under the legacy schema first.
-      const validateLegacy = new Ajv().compile(configSchema0_1_0);
-      const validLegacy = validateLegacy(upgradedConfig);
-      if (!validLegacy) {
-        const failureReason = JSON.stringify(validateLegacy.errors, null, 2);
-        return [{
-          title: 'Config validation failed',
-          preformatted: failureReason,
-        }, false];
-      }
-      // Upgrade from v0.1.0 to v1.0.1 before v1.0.1 schema validation.
-      upgradedConfig = upgrade('0.1.0', upgradedConfig);
+    const [upgradedConfig, upgradeSuccess] = upgradeAndValidate(config);
+    if (upgradeSuccess) {
+      // Initialize the view config according to the initStrategy.
+      const initializedConfig = initialize(upgradedConfig);
+      return [initializedConfig, true];
     }
-    if (upgradedConfig.version === '1.0.0') {
-      // Validate under the legacy schema first.
-      const validateLegacy = new Ajv()
-        .addSchema(cellSetsSchema).addSchema(rasterSchema).compile(configSchema1_0_0);
-      const validLegacy = validateLegacy(upgradedConfig);
-      if (!validLegacy) {
-        const failureReason = JSON.stringify(validateLegacy.errors, null, 2);
-        return [{
-          title: 'Config validation failed',
-          preformatted: failureReason,
-        }, false];
-      }
-      // Upgrade from v1.0.0 to v1.0.1 before v1.0.1 schema validation.
-      upgradedConfig = upgrade('1.0.0', upgradedConfig);
-    }
-    // NOTE: Remove when a view config viewer/editor is available in UI.
-    console.groupCollapsed(`🚄 Vitessce (${packageJson.version}) view configuration`);
-    console.info(`data:,${JSON.stringify(upgradedConfig)}`);
-    console.info(JSON.stringify(upgradedConfig, null, 2));
-    console.groupEnd();
-    const validate = new Ajv()
-      .addSchema(cellSetsSchema).addSchema(rasterSchema).compile(configSchema);
-    const valid = validate(upgradedConfig);
-
-    if (!valid) {
-      const failureReason = JSON.stringify(validate.errors, null, 2);
-      return [{
-        title: 'Config validation failed',
-        preformatted: failureReason,
-      }, false];
-    }
-    // Initialize the view config according to the initStrategy.
-    const initializedConfig = initialize(upgradedConfig);
-    return [initializedConfig, true];
+    return [upgradedConfig, false];
   }, [config]);
 
   // Emit the upgraded/initialized view config
