+++ src/app/view-config-utils.js
@@ -1,14 +1,13 @@
 /* eslint-disable no-plusplus */
-/* eslint-disable camelcase */
-import uuidv4 from 'uuid/v4';
-import { getNextScope, capitalize } from '../utils';
+import packageJson from '../../package.json';
+import { getNextScope } from '../utils';
 import {
   COORDINATION_TYPES,
   DEFAULT_COORDINATION_VALUES,
   COMPONENT_COORDINATION_TYPES,
   AUTO_INDEPENDENT_COORDINATION_TYPES,
 } from './state/coordination';
-
+import { SCHEMA_HANDLERS } from './view-config-versions';
 
 /**
  * Get a list of all unique scope names for a
@@ -168,209 +167,45 @@ export function initialize(config) {
   return config;
 }
 
-/**
- * A helper function for the `upgrade()` function,
- * which helps convert `props.view` (for scatterplot and spatial),
- * into new coordination scopes, setting their values
- * in the coordination space and returning the new scope mappings.
- * This function does mutate the `coordinationSpace` parameter.
- * @param {string} prefix The coordination type prefix,
- * either 'embedding' or 'spatial'.
- * @param {object} view The view prop object containing
- * the properties `.target` and `.zoom`.
- * @param {object} coordinationSpace The coordination space.
- * @returns {object} The new coordination scope names.
- */
-function upgradeReplaceViewProp(prefix, view, coordinationSpace) {
-  const prevZScopes = Object.keys(coordinationSpace[`${prefix}Zoom`]);
-  const prevTXScopes = Object.keys(coordinationSpace[`${prefix}TargetX`]);
-  const prevTYScopes = Object.keys(coordinationSpace[`${prefix}TargetY`]);
+export function upgradeAndValidate(oldConfig) {
+  // oldConfig object must have a `version` property.
+  let nextConfig = oldConfig;
+  let fromVersion;
+  let upgradeFunction; let
+    validateFunction;
 
-  const nextZScope = getNextScope(prevZScopes);
-  const nextTXScope = getNextScope(prevTXScopes);
-  const nextTYScope = getNextScope(prevTYScopes);
+  do {
+    fromVersion = nextConfig.version;
 
-  const { zoom, target: [targetX, targetY] } = view;
-  // eslint-disable-next-line no-param-reassign
-  coordinationSpace[`${prefix}Zoom`][nextZScope] = zoom;
-  // eslint-disable-next-line no-param-reassign
-  coordinationSpace[`${prefix}TargetX`][nextTXScope] = targetX;
-  // eslint-disable-next-line no-param-reassign
-  coordinationSpace[`${prefix}TargetY`][nextTYScope] = targetY;
-  return {
-    [`${prefix}Zoom`]: nextZScope,
-    [`${prefix}TargetX`]: nextTXScope,
-    [`${prefix}TargetY`]: nextTYScope,
-  };
-}
-
-/**
- * Convert an older view config to a newer view config.
- * @param {object} config A v0.1.0 "legacy" view config.
- * @returns {object} A v1.0.0 "upgraded" view config.
- */
-export function upgradeFrom0_1_0(config, datasetUid = null) {
-  const coordinationSpace = {
-    embeddingType: {},
-    embeddingZoom: {},
-    embeddingTargetX: {},
-    embeddingTargetY: {},
-    spatialZoom: {},
-    spatialTargetX: {},
-    spatialTargetY: {},
-  };
-
-  const layout = [];
-  config.staticLayout.forEach((componentDef) => {
-    let newComponentDef = {
-      ...componentDef,
-      coordinationScopes: {},
-    };
-    if (componentDef.component === 'scatterplot') {
-      // Need to set up the coordinationSpace
-      // with embeddingType to replace scatterplot
-      // component prop "mapping".
-      if (componentDef.props.mapping) {
-        coordinationSpace.embeddingType[componentDef.props.mapping] = componentDef.props.mapping;
-        newComponentDef = {
-          ...newComponentDef,
-          coordinationScopes: {
-            ...newComponentDef.coordinationScopes,
-            embeddingType: componentDef.props.mapping,
-          },
-        };
-      }
-      // Need to set up the coordinationSpace
-      // with embeddingZoom / embeddingTargetX/Y to replace scatterplot
-      // component prop "view" ({ zoom, target }).
-      if (componentDef.props.view) {
-        // Note that the below function does mutate the coordinationSpace param.
-        const newScopeValues = upgradeReplaceViewProp(
-          'embedding', componentDef.props.view, coordinationSpace,
-        );
-        newComponentDef = {
-          ...newComponentDef,
-          coordinationScopes: {
-            ...newComponentDef.coordinationScopes,
-            ...newScopeValues,
-          },
-        };
-      }
-    }
-    if (componentDef.component === 'spatial') {
-      // Need to set up the coordinationSpace
-      // with spatialZoom / spatialTargetX/Y to replace spatial
-      // component prop "view" ({ zoom, target }).
-      if (componentDef?.props?.view) {
-        // Note that the below function does mutate the coordinationSpace param.
-        const newScopeValues = upgradeReplaceViewProp(
-          'spatial', componentDef.props.view, coordinationSpace,
-        );
-        newComponentDef = {
-          ...newComponentDef,
-          coordinationScopes: {
-            ...newComponentDef.coordinationScopes,
-            ...newScopeValues,
-          },
-        };
-      }
-    }
-    layout.push(newComponentDef);
-  });
-
-  // Use a random dataset ID when initializing automatically,
-  // so that it changes with each new v0.1.0 view config.
-  // However, check if the `datasetUid` parameter was passed,
-  // which allows for unit testing.
-  const newDatasetUid = datasetUid || uuidv4();
-
-  return {
-    version: '1.0.1',
-    name: config.name,
-    description: config.description,
-    public: config.public,
-    datasets: [
-      {
-        uid: newDatasetUid,
-        name: newDatasetUid,
-        files: config.layers.map(layer => ({
-          type: layer.type.toLowerCase(),
-          fileType: layer.fileType,
-          url: layer.url,
-        })),
-      },
-    ],
-    initStrategy: 'auto',
-    coordinationSpace,
-    layout,
-  };
-}
-
-export function upgradeFrom1_0_0(config) {
-  const coordinationSpace = { ...config.coordinationSpace };
-
-  function replaceLayerType(layerType) {
-    // Layer type could be one of a few things, bitmask or raster at the moment.
-    const isRaster = layerType === 'raster';
-    coordinationSpace[`spatial${capitalize(layerType)}Layer${isRaster ? 's' : ''}`] = {};
-    Object.entries(coordinationSpace.spatialLayers).forEach(([scope, layers]) => {
-      if (Array.isArray(layers) && layers.find(layer => layer.type === layerType)) {
-        const typedLayers = layers
-          .filter(layer => layer.type === layerType)
-          .map((layer) => {
-            const newLayer = { ...layer };
-            delete newLayer.type;
-            return newLayer;
-          });
-        coordinationSpace[`spatial${capitalize(layerType)}Layer${isRaster ? 's' : ''}`][scope] = isRaster ? typedLayers : typedLayers[0];
-      } else {
-        coordinationSpace[`spatial${capitalize(layerType)}Layer${isRaster ? 's' : ''}`][scope] = null;
-      }
-    });
-  }
-
-  if (coordinationSpace.spatialLayers) {
-    replaceLayerType('raster');
-    replaceLayerType('cells');
-    replaceLayerType('molecules');
-    replaceLayerType('neighborhoods');
-    delete coordinationSpace.spatialLayers;
-  }
-
-  const layout = config.layout.map((component) => {
-    const newComponent = { ...component };
-
-    function replaceCoordinationScope(layerType) {
-      const isRaster = layerType === 'raster';
-      if (COMPONENT_COORDINATION_TYPES[newComponent.component].includes(`spatial${capitalize(layerType)}Layer${isRaster ? 's' : ''}`)) {
-        newComponent.coordinationScopes[`spatial${capitalize(layerType)}Layer${isRaster ? 's' : ''}`] = newComponent.coordinationScopes.spatialLayers;
-      }
+    if (!Object.keys(SCHEMA_HANDLERS).includes(fromVersion)) {
+      return [{
+        title: 'Config validation failed',
+        preformatted: 'Unknown config version.',
+      }, false];
     }
 
-    if (newComponent.coordinationScopes && newComponent.coordinationScopes.spatialLayers) {
-      replaceCoordinationScope('raster');
-      replaceCoordinationScope('cells');
-      replaceCoordinationScope('molecules');
-      replaceCoordinationScope('neighborhoods');
-      delete newComponent.coordinationScopes.spatialLayers;
+    [validateFunction, upgradeFunction] = SCHEMA_HANDLERS[fromVersion];
+
+    // Validate under the legacy schema before upgrading.
+    const validLegacy = validateFunction(nextConfig);
+    if (!validLegacy) {
+      const failureReason = JSON.stringify(validateFunction.errors, null, 2);
+      return [{
+        title: 'Config validation failed',
+        preformatted: failureReason,
+      }, false];
     }
-    return newComponent;
-  });
 
-  return {
-    ...config,
-    coordinationSpace,
-    layout,
-    version: '1.0.1',
-  };
-}
+    if (upgradeFunction) {
+      nextConfig = upgradeFunction(nextConfig);
+    }
+  } while (upgradeFunction);
 
-export function upgrade(fromVersion, oldConfig) {
-  if (fromVersion === '0.1.0') {
-    return upgradeFrom0_1_0(oldConfig);
-  }
-  if (fromVersion === '1.0.0') {
-    return upgradeFrom1_0_0(oldConfig);
-  }
-  throw new Error(`Unable to upgrade from unknown version ${fromVersion}`);
+  // NOTE: Remove when a view config viewer/editor is available in UI.
+  console.groupCollapsed(`🚄 Vitessce (${packageJson.version}) view configuration`);
+  console.info(`data:,${JSON.stringify(nextConfig)}`);
+  console.info(JSON.stringify(nextConfig, null, 2));
+  console.groupEnd();
+
+  return [nextConfig, true];
 }
