+++ src/schemas/fixtures/config-1.0.2.bad.json
@@ -1,0 +1,1 @@
+[]
