+++ src/schemas/fixtures/config-1.0.2.bad.message.json
@@ -1,0 +1,11 @@
+[
+  {
+    "dataPath": "",
+    "keyword": "type",
+    "message": "should be object",
+    "params": {
+      "type": "object"
+    },
+    "schemaPath": "#/type"
+  }
+]
