+++ src/schemas/fixtures/config-1.0.2.good.json
@@ -1,0 +1,38 @@
+{
+    "name": "fake view config",
+    "description": "Spatial organization of the somatosensory cortex",
+    "datasets": [
+      {
+        "uid": "fake-dataset",
+        "name": "Fake Dataset",
+        "files": [
+          {
+            "type": "fake",
+            "fileType": "fake.json",
+            "url": "http://example.com",
+            "requestInit": {
+              "method": "POST",
+              "headers": {"x-foo": "bar"},
+              "body": "Body only valid with POSTs.",
+              "mode": "cors",
+              "credentials": "omit",
+              "cache": "no-store",
+              "redirect": "error",
+              "referrer": "foo-bar",
+              "integrity": "foo"
+            }
+          }
+        ]
+      }
+    ],
+    "coordinationSpace": {
+  
+    },
+    "layout": [
+      { "component": "FooBar", "x": 0, "y": 0, "w": 1, "h": 1  },
+      { "component": "FooBar", "props": { "foo": "bar" }, "x": 1, "y": 1, "w": 1, "h": 1 }
+    ],
+    "initStrategy": "auto",
+    "version": "1.0.2"
+  }
+
